refactor: add save participant hook
