+++ frontend/src/hooks/index.js
@@ -2,3 +2,4 @@ export * from "./useRequest";
 export * from "./useToast";
 export * from "./useForm";
 export * from "./useDeleteParticipant";
+export * from "./useSaveParticipant";
